Fix libc portability assumptions
The build assumed several glibc-specific interfaces: the GNU strerror_r return type, nftw action constants, and a non-portable static error-checking mutex initializer. The fanotify configure probe also checked a different header from the one used by the source.

Force the string utility translation unit to select and compile-check the portable XSI strerror_r interface while leaving GNU extensions enabled elsewhere, and redirect existing logging uses through that boundary centrally. Provide the GNU nftw continue and stop constants with their glibc values when libc exposes only the POSIX nftw interface. Fall back to the portable mutex initializer and probe the userspace fanotify header. Make related formatting portable.

+++ src/daemon/fapolicyd.c
@@ -1470,7 +1470,8 @@ int main(int argc, const char *argv[])
 		msg(LOG_WARNING, "Can't increase file number rlimit - %s",
 		    strerror(errno));
 	else
-		msg(LOG_INFO,"Can handle %lu file descriptors", limit.rlim_cur);
+		msg(LOG_INFO, "Can handle %llu file descriptors",
+		    (unsigned long long)limit.rlim_cur);
 
 	// get more time slices because everything is waiting on us
 	errno = 0;
